Move the mobile navbar in a dedicated component
