+++ src/routes/__layout.svelte
@@ -2,7 +2,7 @@
   import '../app.css';
   import 'bootstrap-icons/font/bootstrap-icons.css';
   import Footer from '$lib/components/footer.svelte';
-  import Navbar from '$lib/components/navbar.svelte';
+  import Navbar from '$lib/components/navbar/navbar.svelte';
 </script>
 
 <div class="flex flex-col min-h-screen dark:bg-gray-900">
